Locates R symbols by parsing the ELF file instead of using dlopen().
It turns out to be relatively straightforward to parse symbol tables in
ELF shared libraries directly, although it may make it more difficult to
port to other platforms in the short term. This approach is also less of
a hack, and should make it easier to handle distributions of R that do
not link to a shared libR.so (e.g. not built with --enable-R-shlib),
since we can parse the executable ELF instead (#7).

This patch does introduce a dependency on libelf, but this is installed
on most Linux distributions by default.

It also removes a blocker for profiling R processes running inside a
(Docker) container.

Closes #6.