Removes unnecessary heap memory usage when copying remote data.
Prior to this commit, most structures were repeatedly heap allocated,
which is not necessary. All copy_*() functions have now moved the
responsibility for allocation to the caller, which allows us to use the
stack for most cases instead.

This dramatically reduces memory allocation when running the
application. According to valgrind, we now only allocate only 9 times
regardless of runtime. Before, we could allocate many orders of
magnitude more -- for example, at 49 Hz for 5 seconds I observed 10,128
allocations when profiling an idle Shiny app.

Obviously, this should improve the performance of the profiler.