Refactor pid_t-related functionality into a new process handle API.
Processes are not pid_t on Windows, so we need a new type for them, and
since it is possible for them to have "state", we might as well create
an new internal API for these "process handles".

Handles are now used when reading memory, when looking up R's symbols,
and when doing system-specific things like "suspending the process".

This should make these APIs possible to implement for Windows (#3).