windows: Fixes the definition of JMP_BUF.
On POSIX, sigjmp_buf is identical to jmp_buf. But it turns out that R
has its own implementation of setjmp for Windows that defines a new
structure.

With this change, RCNTXT now matches the size of R's implementation on
that platform and the call SEXP can be correctly identified.