macos: Initial macOS port.
This works much the same as the Linux implementation, querying the
process for loaded Mach-O libraries, choosing one that looks like R, and
then reading the symbol table in that binary to get the corresponding
addresses.

Note that this initial implementation *does not suspend the process*,
which may yield unpredictable results.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>